transit cache is an Interface implemented by wrapped versions of syncmap and golang-lru
removes Purge() from Cache interface. Adds Size() to Cache interface. Adds a RWMutex to LockManager.

bubbles up lru.New2Q(). path-config description rewording. use %q instead of %s for an Sprintf ErrorResponse that is shown to users

use errwrap to wrap errors that bubble up to the user

use Operations instead of Callbacks

added summaries to operations

use the term unlimited instead of syncmap for an unlimited cache that is
backed by a syncmap

if a user specifies a cache-type and cache-size that is already active
this is now a no op instead of resetting the cache

use underscore separators in API response

user doesn't need to know anything about the underlying cache type anymore

remove unused Len function from implementors of Cache interface

NewLockManager initializes it s own cache instead of defering this to the transit backend - this is because other things call NewLockManager and expect a cache to be available right away